Make the p elements a bit lighter
I started this amazing course a few days ago, but something that's bugging me out for some time is that the text is a bit too gray for me, and with the background being almost black, is not that comfortable to read after a while.
So I propose to make the font-color a little brighter, changing it from #b3b3b3 to #c8c8c8. Is not a big change, but I feel is more comfortable to read.